\encoding{UTF-8}
\name{Ausmale}
\alias{Ausmale}
\title{Australian Male Mortality Rates}
\description{
The data consist of set of measurements across years of male mortality rates 
in Australia from package \code{fds}.
}
\usage{Ausmale}
\format{A list with years in the first component and a 101 times 103 matrix, 
years in rows and cohorts in columns, in the second component.}
\source{fds}
\keyword{datasets}
