\encoding{UTF-8}
\name{merval}
\alias{merval}
\title{Merval Index}
\description{
  The data consist of an low and high values of the Merval Index stock market 
  from Argentina; the data were gathered from Yahoo Finance.
}
\usage{merval}
\format{A dataframe with 5269 observations with daily minimum, maximum, 
open and close index values.}
\source{Yahoo Finance}
\keyword{datasets}
